requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
jinja2==3.1.2
python-multipart==0.0.6
alembic==1.12.1
python-dotenv==1.0.0
pyjwt==2.8.0